# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/collections": {
      "get": {
        "operationId": "CollectionsController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page, from 1 to 100.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "integer"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "The `nextCursor` from the previous page. Opaque.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "A case-insensitive substring match, trimmed. `%`, `_` and `\\` match literally.",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionPageDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON, an unknown field, or an invalid query parameter",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "The Bearer token is missing, invalid or expired",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lists the caller's Collections by name, ignoring case.",
        "tags": [
          "Collections"
        ]
      },
      "post": {
        "operationId": "CollectionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON, an unknown field, or an invalid query parameter",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "The Bearer token is missing, invalid or expired",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "409": {
            "description": "The Owner already has a Collection with this name, ignoring case",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "415": {
            "description": "The body isn't application/json",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "422": {
            "description": "A value is invalid. Each invalid field is listed with a JSON Pointer.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Collections"
        ]
      }
    },
    "/collections/{id}": {
      "get": {
        "operationId": "CollectionsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionDto"
                }
              }
            }
          },
          "401": {
            "description": "The Bearer token is missing, invalid or expired",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "Doesn't exist, belongs to another Owner, or has a malformed ID. Always the same body.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Collections"
        ]
      },
      "put": {
        "operationId": "CollectionsController_replace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceCollectionBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON, an unknown field, or an invalid query parameter",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "The Bearer token is missing, invalid or expired",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "Doesn't exist, belongs to another Owner, or has a malformed ID. Always the same body.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "409": {
            "description": "The Owner already has a Collection with this name, ignoring case",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "415": {
            "description": "The body isn't application/json",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "422": {
            "description": "A value is invalid. Each invalid field is listed with a JSON Pointer.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Collections"
        ]
      },
      "patch": {
        "operationId": "CollectionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCollectionBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON, an unknown field, or an invalid query parameter",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "The Bearer token is missing, invalid or expired",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "Doesn't exist, belongs to another Owner, or has a malformed ID. Always the same body.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "409": {
            "description": "The Owner already has a Collection with this name, ignoring case",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "415": {
            "description": "The body isn't application/json",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "422": {
            "description": "A value is invalid. Each invalid field is listed with a JSON Pointer.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Behaves like PUT, because `name` is the only writable field.",
        "tags": [
          "Collections"
        ]
      },
      "delete": {
        "operationId": "CollectionsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "The Bearer token is missing, invalid or expired",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "Doesn't exist, belongs to another Owner, or has a malformed ID. Always the same body.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deletes the Collection. Its Bookmarks are kept, and become Uncategorised.",
        "tags": [
          "Collections"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bookmark manager API",
    "description": "The contract is described in API_DESIGN.md. Every route needs a Bearer token.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ProblemDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "number"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ]
      },
      "FieldErrorDto": {
        "type": "object",
        "properties": {
          "pointer": {
            "type": "string",
            "description": "A JSON Pointer into the request body."
          },
          "detail": {
            "type": "string"
          }
        },
        "required": [
          "pointer",
          "detail"
        ]
      },
      "ValidationProblemDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "number"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldErrorDto"
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "errors"
        ]
      },
      "CollectionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "bookmarkCount": {
            "type": "number",
            "description": "How many of the Owner's Bookmarks are in this Collection."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "bookmarkCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "CollectionPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CollectionDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "description": "Pass as `cursor` to get the next page. Null on the last page."
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
      "CreateCollectionBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Trimmed. Unique per Owner, ignoring case.",
            "example": "Reading",
            "minLength": 1,
            "maxLength": 100
          }
        },
        "required": [
          "name"
        ]
      },
      "ReplaceCollectionBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Trimmed. Unique per Owner, ignoring case.",
            "example": "Reading",
            "minLength": 1,
            "maxLength": 100
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCollectionBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Trimmed. Unique per Owner, ignoring case. Absent means unchanged.",
            "minLength": 1,
            "maxLength": 100
          }
        }
      },
...
    }
  }
}